Controller/Service update

+++ backend/flightplanner/src/main/java/com/example/test/repository/SeatRepository.java
@@ -3,6 +3,6 @@ package com.example.test.repository;
 import com.example.test.model.Seat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeatRepository extends JpaRepository<Seat, Long> {
+public interface SeatRepository extends JpaRepository<Seat, Integer> {
 
 }
