+++ backend/flightplanner/src/main/java/com/example/test/repository/TicketRepository.java
@@ -3,8 +3,6 @@ package com.example.test.repository;
 import com.example.test.model.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
-public interface TicketRepository extends JpaRepository<Ticket, Long> {
+public interface TicketRepository extends JpaRepository<Ticket, Integer> {
 
 }
